extract system info logging into its own file
