+++ src/logging/mod.rs
@@ -1,9 +1,9 @@
 mod coredump;
 mod error;
+mod sysinfo;
 
 use std::{
     self,
-    collections::BTreeMap,
     fs::File,
     panic::{PanicHookInfo, set_hook},
     path::Path,
@@ -11,7 +11,7 @@ use std::{
 };
 
 pub use error::{ErrorContext, ErrorInfo, ErrorSeverity, ErrorWithInfo, crash_and_burn};
-use tracing::{Level, info, warn};
+use tracing::{Level, info};
 use tracing_subscriber::{EnvFilter, fmt::format::FmtSpan};
 
 use crate::logging::error::RemediationAdvice;
@@ -64,10 +64,10 @@ pub fn init_logging_parent(paths: &LogPaths) -> Arc<error::ErrorContext> {
 
     init_tracing_subscriber(file);
 
-    log_program_info();
-    log_uname();
-    log_info_files();
-    log_environment_variables();
+    sysinfo::log_program_info();
+    sysinfo::log_uname();
+    sysinfo::log_info_files();
+    sysinfo::log_environment_variables();
 
     let coredump_instructions = self::coredump::instructions();
     info!("Guessed system coredump handler to be {coredump_instructions:?}");
@@ -123,82 +123,4 @@ fn init_tracing_subscriber(file: File) {
         .with_file(true)
         .with_line_number(true)
         .init();
-}
-
-fn log_program_info() {
-    let name = env!("CARGO_PKG_NAME");
-    let version = env!("CARGO_PKG_VERSION");
-    tracing::info!("Starting {name} v{version}");
-}
-
-pub fn log_uname() {
-    let Ok(uname) = uname::uname().inspect_err(|e| warn!("Unable to get uname info: {e}")) else {
-        return;
-    };
-    info!(?uname.sysname, ?uname.version, ?uname.release, ?uname.machine, "uname info");
-}
-
-/// Log the contents of some files containing useful debug info about the user's
-/// system
-fn log_info_files() {
-    const INFO_FILES: &[&str] = &[
-        "/etc/os-release",
-        "/etc/lsb-release",
-        "/etc/redhat_release",
-        "/etc/debian_version",
-        "/System/Library/CoreServices/SystemVersion.plist",
-        "/sys/class/dmi/id/product_name",
-    ];
-
-    for path in INFO_FILES {
-        match std::fs::read_to_string(path) {
-            Ok(contents) => info!("{path} contents:\n{}", contents.trim()),
-            Err(e) => match e.kind() {
-                std::io::ErrorKind::NotFound => info!("{path} not found, skipping"),
-                _ => warn!("Failed to read {path}: {e}"),
-            },
-        }
-    }
-}
-
-fn log_environment_variables() {
-    const ENVIRONMENT_VARIABLES: &[(&str, &[&str])] = &[
-        (
-            "Terminal emulator",
-            &[
-                "TERM",
-                "TERM_PROGRAM",
-                "TERM_PROGRAM_VERSION",
-                "VTE_VERSION",
-                "GNOME_TERMINAL_SCREEN",
-            ],
-        ),
-        (
-            "Color capabilities",
-            &[
-                "COLORTERM",
-                "FORCE_COLOR",
-                "NO_COLOR",
-                "CLICOLOR",
-                "CLICOLOR_FORCE",
-            ],
-        ),
-        ("Multiplexer", &["TMUX", "TMUX_PANE", "SCREEN", "STY"]),
-        ("Locale", &["LANG", "LC_ALL"]),
-    ];
-
-    let vars = ENVIRONMENT_VARIABLES
-        .iter()
-        .map(|(k, varnames)| {
-            (
-                *k,
-                varnames
-                    .iter()
-                    .map(|var| (*var, std::env::var(var).ok()))
-                    .collect::<BTreeMap<&'static str, Option<String>>>(),
-            )
-        })
-        .collect::<BTreeMap<&'static str, _>>();
-
-    tracing::info!("Environment variables detected: {vars:#?}")
 }
